RDKBDEV-599: GPON Manager: Use InterfaceName DM parameter instead of InterDomainName
Reason for change: Device.X_RDK_ONT.Veip.1.InterDomainName value returning as NULL.
Value of this parameter is essential for RdkGponManager state machine to operate as expected.
This parameter is used for the integration with the Ethernet Agent.
So changed to Device.X_RDK_ONT.Veip.1.InterfaceName (which is veip0).

Test Procedure:
        1. Check the value of parameter Device.X_RDK_ONT.Veip.1.InterfaceName
        2. Ensure that GPON Manager state machine is getting started and erouter0 interface got an IPAddress.
        3. Device comes online.
        Attached in Jira RDKBDEV-599.

Risks: Low.

Signed-off-by: Abhishek Mishra <abhishek.mishra@external.telekom.de>
Change-Id: Ie838515d076d0ad504d7da93674c284eda573f9a